add enroll setting

+++ client/src/components/Enroll.js
@@ -7,17 +7,39 @@ const Enroll = ({ currentUser, setCurrentUser }) => {
   let [courseData, setCourseData] = useState(null);
   let [foundUrl, setFoundUrl] = useState(null);
 
+  const enrollCourse = (e) => {
+    if (currentUser.user.role === "student") {
+      CourseService.enroll(e.target.id, currentUser.user._id)
+        .then((item) => {
+          window.alert(item.data);
+          navigate("/profile");
+        })
+        .catch((err) => {
+          console.log(err);
+        });
+    } else {
+      window.alert("只有學生才能註冊喔~");
+    }
+  };
+
   const deleteCourse = (e) => {
     CourseService.getCourseWithID(e.target.id).then((item) => {
       if (item.data.instructor._id === currentUser.user._id) {
-        CourseService.deleteCourse(e.target.id)
-          .then(() => {
-            window.alert("課程已經被刪除，現在幫你導到搜尋課程的頁面");
-            navigate("/search");
-          })
-          .catch((err) => {
-            console.log(err);
-          });
+        let check = window.prompt(
+          `請輸入該課程的名稱「${item.data.title}」作為確認`
+        );
+        if (check === item.data.title) {
+          CourseService.deleteCourse(e.target.id)
+            .then(() => {
+              window.alert("課程已經被刪除，現在幫你導到搜尋課程的頁面");
+              navigate("/search");
+            })
+            .catch((err) => {
+              console.log(err);
+            });
+        } else {
+          window.alert("輸入的與課程名稱不同，請重新操作");
+        }
       } else {
         window.alert("只有該講師才能刪除課程");
       }
@@ -60,7 +82,9 @@ const Enroll = ({ currentUser, setCurrentUser }) => {
         ></iframe>
       )}
       {currentUser && currentUser.user.role === "student" && (
-        <button>註冊課程</button>
+        <button id={currentUser.user.course} onClick={enrollCourse}>
+          註冊課程
+        </button>
       )}
       {currentUser && currentUser.user.role === "instructor" && (
         <button id={currentUser.user.course} onClick={deleteCourse}>
